Call check_for_rustc_errors_attr from start_codegen
